Project/Trabalho1/react_page/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.1
transformers==4.35.2
requests==2.31.0
pydantic==2.4.2